requirements.txt
django
gunicorn
whitenoise
django-cors-headers
djangorestframework
drf-spectacular
python-dotenv
pillow